# docs/swagger.yaml
definitions:
  request.CreateMessageRequest:
    properties:
      content:
        maxLength: 255
        type: string
      phone_number:
        maxLength: 20
        type: string
    required:
    - content
    - phone_number
    type: object
  response.Response:
    properties:
      data: {}
      message:
        type: string
      status:
        type: string
    type: object
info:
  contact: {}
paths:
  /message:
    post:
      description: Create and send a new message
      parameters:
      - description: Create message request
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/request.CreateMessageRequest'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Create a new message
      tags:
      - messages
  /send-messages:
    post:
      description: Start or stop automatic message sending based on the query parameter
        'action'.
      parameters:
      - description: Action to perform (start or stop)
        in: query
        name: action
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Start or stop automatic message sending
      tags:
      - messages
  /sent-messages:
    get:
      description: Retrieve all sent messages from the system
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get sent messages
      tags:
      - messages
swagger: "2.0"